refactor getClosureInputsAndOutputs to respect dependency order (#14)
Refactor getClosureInputsAndOutputs to correctly handle out-of-order statements (i.e. forward references). This is a continuation of APPS-247 - I can either merge this with the bugs/APPS-247 first, or wait until that one is merged and then change this PR to be against main.

This highlights the issue that forward references are currently not handled during type inference. I have added an issue (APPS-273) to address this later.