Null array as empty (#48)
DNAnexus does not distinguish between null and empty fo array inputs. Currently, if an empty array is passed to a non-optional WDL array input there is an error, even if the WDL array is allowed to be empty. This change (which depends on a new version of wdlTools) substitutes an empty array for a null value in the case of a non-optional array that is allowed to be empty.