Fix APPS-369 and APPS-370 (#56)
Fixes two bugs:

APPS-369 - structs out of order in generated code. This bug was due to an unnecessary conversion of a Vector to Map in CodeGenerator, which disordered the items.
APPS-370 - unable to resolve nested field references. There were two issues here:
Unnecessarily including outputs that are direct pass-throughs of inputs in the closure when determining the inputs to an Output stage.
A bug in the resolution of deeply nested field references in wdlTools.eval.Eval that is fixed in 0.11.18.