APPS 384 - fix expression evaluation errors (#59)
Fixes two errors uncovered by test case in APPS-384:

expression evaluator was not handling values wrapped in V_Optional
compound references were not handled correctly - a reference like row.left in an expression would only be evaluated if the name "row.left" were in the evaluation context. Now, if it is missing (say "row" is in the evaluation context, but not "row.left"), row.left is turned into an expression and evaluated using the expression evaluator.
Updates to wdlTools 0.12.3. This update also enforces uniqueness of variable names within the same scope.