APPS-381 Add ECR support (#63)
Adds a new option, awsRegion, to extras, and makes username optional and mutually exclusive. If awsRegion is specified, then the registry is assumed to be ECR and a different docker script is used when compiling applications that handles installing the aws cli and logging in to docker using the AWS-specific method.

Also refactors Extras to use Spray JSON protocols for cleaner parsing code.