APPS-395 - optionally use separate output folder per call (#66)
Currently, all output files for an analysis are placed in the same output folder. This can lead to name collisions, e.g. when a scatter job produces a file with a generic name, and then a downstream task tries to localize an array of those files (which all have the same name). This PR adds an option (-separateOutputs) to put the output files from each job in a separate folder.

Also updates to dxApi 0.1.11.