APPS-422 - handle block inputs that are compound reference to private variables (#76)
If a WdlBlock contains a declaration of an object or struct type, it will yield a WdlBlockInput with a compound name. However, when this is evaluated, the environment may only contain a nested component of the name, so it results in a missing required input. This PR fixes this scenario by searching for the fully qualified name and all possible sub-names in the environment.