APPS-288 - fixes issue with using both streaming and non-streaming files (#79)
The issue was in the generated job script in the bash code block for dxda - it cd's to / at the beginning and then is supposed to cd back to the home directory at the end, but instead it was cd-ing to the work dir. This only causes an issue if dxfuse is also used, because it tries to source environment, which is in the home dir. The error was not showing up in the log because we pipe the output from source environment to /dev/null to avoid leaking the token to the logs, so it was also swallowing the error message.

Also fixes the parent directory name written to the dxfuse manifest (removes the extra leading '/').

Also updates to dxCommon 0.2.10 which fixes another issue - the LocalizationDisambiguator was creating directories in the dxfuse mount point, which is wrong behavior, but I guess dxfuse doesn't complain as long as the directory is empty.